fixed bug in rename_folder()
this bug occurred because os.listdir listed the .py file first (whether the .py file gets listed first will depend on the names of the other files in the same trajectory, e.g. if other  file name starts with letter "z" which means they will be listed after "rename_parser_script.py" which starts with "r" (alphabetical order r comes before z)), which caused downstream of the function taking .py as the extension, of which there is only the rename_parser_script.py itself and none of the md files will be added to file_list.